refactor: move probe send notification function
